requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
psutil>=5.9.0
requests>=2.31.0
eventlet>=0.33.0
python-engineio>=4.8.0
